requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud>=0.34.0
google-cloud-datastore>=1.9.0
google-cloud-storage>=1.19.0
google-cloud-logging>=1.12.1
flask-cors>=3.0.8
